wip: partials variables spec passing

+++ spec/spec_helper.rb
@@ -40,11 +40,14 @@ module Helper
   # Copies spec/fixtures/lono_project to tmp/lono_project,
   # Main fixture we'll use because it's faster
   def copy_lono_project
+    destroy_lono_project # just in case KEEP_TMP_PROJECT is used
     FileUtils.cp_r("spec/fixtures/lono_project", "tmp/lono_project")
   end
 
   def destroy_lono_project
-    FileUtils.rm_rf(Lono.root)
+    # Only use KEEP_TMP_PROJECT if you are testing exactly 1 spec for debugging
+    # or it'll affect other tests.
+    FileUtils.rm_rf(Lono.root) unless ENV['KEEP_TMP_PROJECT']
   end
 end
